Fixes the response of LRU.XGET.
The extended get may return stale cache entries but signals this by
setting the alive to 0 and refresh to 1. However, for a certain period
all successive calls will receive a 0 for refresh to prevent that multiple
callers try to compute the same value at once.

In this case however, we have to toggle to alive flag to 1 as otherwise
the result looks like a fully stale (unusable) cache entry.