Adds a consistent restart hook.
If a panic is raised while handling a command, this will
most probably crash one of the tokio threads leaving the
remaining parts in an inconsistent state.

Sadly, the server still looks to be healthy - but will not
actually execute any incoming command.

We therefore install a custom handler, which crashes
the whole process once a panic is detected. This way
an external watchdog like docker-compose can restart
a new, sane and consistent process.