Implements Sets for InfoGraphDB.
Next to raw data tables / code lists, we often need all kinds of lists
or sets which determine which code is accepted in which scenario.

Like "in this part of the app, only these languages codes are allowed"
or "for this field, only accept the following unit codes".

Instead of providing this as extra table, we now add support for simple
sets (lists of strings) which can be checked in a very efficient manner.