Implements a simple way of detecting if the repository background actor is idle or busy.
The foreground actor keeps two counters: an epoch for the foreground and one for the
background. If the REPO.INC_COUNTER command is received, the foreground epoch
is incremented immediately but the background is left untouched. We rather queue a
special background task for this. Once the background task handles this command, it
immediately sends back an appropriate background event, which is then used to finally
increment the background epoch counter within the frontend actor.

This way one can tell that the background actor is idle, as soon as REPO.EPOCHS
returns equal values for foreground and background epochs (after an REPO.INC_EPOCH)
 has been requested.