No longer returns a whole map when performing an i18n call.
If an inner map is detected and neither the requested language
nor its fallback is found as key, we previously returned the whole
map.

For use cases where only partial data is available, it would be
better to return an empty string to indicate that no data is available.

Returning a map is difficult anyway as we never know when a the
user expects a JSON map and when a translated lookup is requested.

Note that for "non translated" calls, we still support returning inner maps
as in this case these is no such contradiction.