Moves common functionality out of Jupiter into Apollo.
We determined that a highly available server component including a
live-reloadable config and proper signal handling might be useful
to other projects. Thus this has been factored out of Jupiter and is
now called "Apollo".

The most significant change is, that the config now operates on Yaml
objects instead of wrapping them as "Doc". As the config is more or
less tiny, this doesn't have much of an effect.