Optimizes queries for inner indices on missing data.
We commonly have tables which contain a "mappings" dict, which contains
additional values used for lookups. For many tables, these mappings are
completely empty or very sparse. Still we perform a lot of
normalizations/lookups and attempt to query these fields anyway
(just to be sure).

Previously, we missed an index in this case (as the inner indices
weren't created without data) and performed a pointless table scan.

We now detect this pattern and simplify this query.