Change the order sync events generated by the repository.
We now first emit all "FileDelete" events and then the "FileChanged"
events. This way, if e.g. a loader is renamed the delete is processed
first -> the referenced data is unloaded. Then the "new" file is
processed and the data will be loaded again.

Previously, we first processed the new loader -> the data was loaded
(kept alive) when we processed the deleted file and the date was
unloaded and gone :(