Makes IDB.LOOKUP and IDB.QUERY case-insensitive (in some cases).
In order for these commands to be case-insensitive, the field being
queried has to have a fulltext index and the search value itself has
to be given as lowercase.

As an example, if a table contains "Austria" in the field name, for
which also an fulltext index is defined, a query like

IDB.LOOKUP table name austria

will match properly.