Adds support for secondary cache keys in the LRU caches.
A secondary key cannot be used for retrieval, but only to remove one or
more entries associated with a given key. This is mainly useful, if
one or more cache entries are based on several objects, like a database
entities.

Now if a central database entity changes, all cache entries associated
with this entity have to be removed from the cache. Therefore these
entries can be associated with one or more secondary keys (e.g. database
IDs) which can then be flushed (using LRU.REMOVES <cache> <key>) once
the entity is modified.