Removes jemalloc as main allocator
We used this to properly track the ram usage. However,
it turned out, that the reported ram usage is incomplete
(e.g. as C libraries still malloc) and also, we could not
use heaptrack to analyze memory usage (as this also
requires using malloc).

Note that dependencies have been updated while
we`re at it (using cargo update)...